Start adding 3d renderer and camera and try not to break existing usage for 2d, for now I am copying the 2d camera and renderer, later these may merge into the same with enum, maybe traits, or soemthing else, depending on how much is the same / different between each
